+++ src/Controller/CharacterController.php
@@ -6,13 +6,19 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Attribute\Route;
 use App\Entity\Character;
+use App\Service\CharacterServiceInterface;
 
 class CharacterController extends AbstractController
 {
-    #[Route('/characters/', name: 'app_character_display', methods: ['GET'])]
-    public function display(): JsonResponse
+
+    public function __construct(
+        private CharacterServiceInterface $characterService
+    ) {}
+
+    #[Route('/characters/', name: 'app_character_create', methods: ['POST'])]
+    public function create(): JsonResponse
     {
-        $character = new Character();
-       return new JsonResponse($character->toArray());
+        $character = $this->characterService->create();
+        return new JsonResponse($character->toArray(), JsonResponse::HTTP_CREATED);
     }
 }