+++ ChangeLog.md
@@ -1,3 +1,21 @@
-ChangeLog
 v0.1
+
+- Core files
+
+v0.2.0
+- Ajout Entity Character
+- Ajout .htaccess
+- Ajout VarDumper (dev)
+- Ajout CharacterController
+- Ajout tests
+- 
+v0.2.1
+- Ajout des methods à la route
+
+v0.3
+- Ajout de la route create character
+- Ajout de la sauvegarde en base de données
+
+
+
 - Core files